vp8/encoder/vp8_quantize: fix int sanitizer warnings
 implicit conversion from type 'int' of value 65536
 (32-bit, signed) to type 'short' changed the value to 0 (16-bit, signed)

 BUG=webm:1615

+++ vp8/encoder/vp8_quantize.c
@@ -175,7 +175,7 @@ static void invert_quant(int improved_quant, short *quant, short *shift,
     *quant = (1 << 16) / d;
     *shift = 0;
     /* use multiplication and constant shift by 16 */
-    *shift = 1 << (16 - *shift);
+    *shift = (short)(1 << (16 - *shift));
   }
 }
